+++ sharedvision.js
@@ -3,23 +3,28 @@ import {initializeSources, onSetShareVision, compatibleCore} from "./src/misc.js
 import {heyWait_onTileHud, setTriggerHappyActive, setHeyWaitActive} from "./src/externalModules.js";
 import {visionConfig} from './src/visionConfig.js';
 import {libWrapper} from './src/shim.js';
-import {registerSettings} from "./src/settings.js";
-import {socketInit, emitSharedVision} from "./src/socket.js";
+import {registerSettings, migrateSettings} from "./src/settings.js";
+import {socketInit, emitSharedVision, updateSight} from "./src/socket.js";
 import {isVisionSourceOverride,updateOcclusionOverride} from './src/overrides.js';
+import { updateToken } from "./src/tokenLayer.js";
 
 export const moduleName = "SharedVision";
 export let midiQOL;
 
-Hooks.once('init', function()                                         { onInit() });
-Hooks.once('setup', function()                                        { onSetup() });
-Hooks.on('getSceneControlButtons', (controls)                     =>  { pushControlButtons(controls) });  //Register control button
-Hooks.on('setShareVision',(data)                                  =>  { onSetShareVision(data) });
-Hooks.on('canvasReady',()                                         =>  { onCanvasReady() });
-Hooks.on('updateToken',()                                         =>  { emitSharedVision(game.settings.get(moduleName,'enable'), false)});
+//CONFIG.debug.hooks = true;
 
-//Hooks.on('controlToken',(token,controlled)                        =>  { triggerHappy_ControlToken(token,controlled) });
-//Hooks.on('preUpdateToken',(scene, embedded, update)               =>  { triggerHappy_onPreUpdateToken(scene, embedded, update) });
-Hooks.on('renderTileHUD', async (tileHud, html)                   =>  { heyWait_onTileHud(tileHud,html) });
+Hooks.once('init', function()                                           { onInit() });
+Hooks.once('setup', function()                                          { onSetup() });
+Hooks.once('ready', function()                                          { onReady() });
+Hooks.on('getSceneControlButtons', (controls)                       =>  { pushControlButtons(controls) });  //Register control button
+Hooks.on('setSharedVision',(data)                                   =>  { onSetShareVision(data) });
+Hooks.on('canvasReady',()                                           =>  { onCanvasReady() });
+Hooks.on('updateToken', (data)                                      =>  { onUpdateToken(data) });
+Hooks.on('sightRefresh', (data)                                     =>  { onSightRefresh(data) });
+Hooks.on('renderTileHUD', async (tileHud, html)                     =>  { heyWait_onTileHud(tileHud,html) });
+Hooks.on('combatStart', ()                                          =>  { onCombat('start') });
+Hooks.on('deleteCombat', ()                                         =>  { onCombat('end') });
+Hooks.on('updateCombat', (a,b)                                      =>  { onUpdateCombat(a,b) });
 
 function onInit(){
     registerSettings(); 
@@ -36,7 +41,7 @@ function onInit(){
                     return game.user.isGM;
                 },
                 callback: (li) => {
-                    const actor = compatibleCore("0.9.0") ? this.constructor.collection.get(li.data("documentId")) : this.constructor.collection.get(li.data("entityId"));
+                    const actor = this.constructor.collection.get(li.data("documentId"));
                     if (actor) {
                         let dialog = new visionConfig();
                         dialog.setActor(actor);
@@ -50,7 +55,6 @@ function onInit(){
     setTimeout(function(){
         updateNotification();
     },500)
-    
 }
 
 function onSetup() {
@@ -58,8 +62,8 @@ function onSetup() {
     setTriggerHappyActive(triggerHappy != undefined && triggerHappy.active == true)
 }
 
-Hooks.once('ready', function() {
-    
+function onReady() {
+    if (game.user.isGM) migrateSettings(); 
 
     const heyWait = game.modules.get("hey-wait");
     setHeyWaitActive(heyWait != undefined && heyWait.active == true)
@@ -71,16 +75,38 @@ Hooks.once('ready', function() {
     
     if(game.modules.get('lib-wrapper')?.active) {
         libWrapper.register("SharedVision", "Token.prototype._isVisionSource", isVisionSourceOverride, "OVERRIDE");
-        if (compatibleCore('0.8.6')) libWrapper.register("SharedVision", "ForegroundLayer.prototype.updateOcclusion", updateOcclusionOverride, "OVERRIDE");
+        libWrapper.register("SharedVision", "ForegroundLayer.prototype.updateOcclusion", updateOcclusionOverride, "OVERRIDE");
     }
         
     else {
         Token.prototype._isVisionSource = isVisionSourceOverride;
-        if (compatibleCore('0.8.6')) ForegroundLayer.prototype.updateOcclusion = updateOcclusionOverride;
+        if (!compatibleCore('10.0')) ForegroundLayer.prototype.updateOcclusion = updateOcclusionOverride;
     }
 
-    if (!game.user.isGM) initializeSources();
-});
+    initializeSources();
+}
+
+let currentlyUpdatingToken;
+let currentlyUpdatingTokenVisible = false;
+let currentlyUpdatingTokenTimer;
+
+function onUpdateToken(data) {
+    emitSharedVision(game.settings.get(moduleName,'enable'), false);
+    const token = compatibleCore('10.0') ? canvas.tokens.placeables.find(t => t.id == data.id) : canvas.tokens.placeables.find(t => t.id == data.data._id);
+    
+    currentlyUpdatingToken = compatibleCore('10.0') ? data.id : data.data._id;
+    currentlyUpdatingTokenVisible = token?.visible;
+    currentlyUpdatingTokenTimer = setTimeout(function(){
+        updateToken(token);
+        updateSight(currentlyUpdatingToken)
+        currentlyUpdatingToken = undefined
+    }, 250);
+}
+
+function onSightRefresh(data) {
+    clearTimeout(currentlyUpdatingTokenTimer)
+    if (currentlyUpdatingToken != undefined && currentlyUpdatingTokenVisible) updateSight(currentlyUpdatingToken)
+}
 
 async function onCanvasReady(){
     if(midiQOL && game.modules.get('lib-wrapper')?.active == false) {
@@ -89,7 +115,26 @@ async function onCanvasReady(){
     
     const enable = game.settings.get(moduleName,'enable');
     if (game.user.isGM) emitSharedVision(enable);
-    else initializeSources();
+    initializeSources();
+}
+
+function onCombat(mode) {
+    if (game.user.isGM == false) return;
+    const combatConfig = game.settings.get(moduleName,'combatConfig');
+    if (Object.keys(combatConfig).length === 0) return;
+    let data = {};
+    let global = combatConfig[mode].global;
+    if (global == 'enable') data.globalSharedVision = true;
+    else if (global == 'disable') data.globalSharedVision = false;
+    let disableAll = combatConfig[mode].disableAll;
+    if (disableAll == 'enable') data.disableAll = true;
+    else if (disableAll == 'disable') data.disableAll = false;
+
+    onSetShareVision(data)
+}
+
+function onUpdateCombat(a,b) {
+    if (a.previous.round == 0 && a.previous.turn == 0) onCombat('start');
 }
 
 function updateNotification(){