+++ src/socket.js
@@ -1,5 +1,6 @@
-import {initializeSources, onSetShareVision} from "./misc.js";
+import {initializeSources, onSetShareVision, revealTokenFog} from "./misc.js";
 import {heyWait_onTrigger} from "./externalModules.js";
+import { moveToken, updateAllTokens } from "./tokenLayer.js";
 
 /*
  * Set up the socket used to communicate between GM and player clients
@@ -7,8 +8,15 @@ import {heyWait_onTrigger} from "./externalModules.js";
 export function socketInit(){
     game.socket.on(`module.SharedVision`, (payload) =>{
         if (game.user == null) return;
-        if (game.user.isGM == false && payload.msgType == "enable") initializeSources(payload.newSource);
+        if (game.user.isGM == false && payload.msgType == "enable") {
+            initializeSources(payload.newSource);
+        }
         else if (game.user.isGM && payload.msgType == "userSet") onSetShareVision({enable:payload.enable});
+        else if (payload.msgType == 'updateSight') {
+            const token = canvas.tokens.placeables.find(t => t.id == payload.tokenId);
+            revealTokenFog(token);
+            moveToken(token);
+        }
     }); 
     
     game.socket.on('module.hey-wait', (payload) =>{
@@ -33,6 +41,15 @@ export function userSet(en){
         "msgType": "userSet",
         "senderId": game.userId, 
         "enable": en
+    };
+    game.socket.emit(`module.SharedVision`, payload);
+}
+
+export function updateSight(tokenId) {
+    const payload = {
+        "msgType": "updateSight",
+        "senderId": game.userId, 
+        "tokenId": tokenId
     };
     game.socket.emit(`module.SharedVision`, payload);
 }